Fix conflict between chunks
