+++ src/extensions/typography/constants.js
@@ -1,1 +1,2 @@
 export const TYPOGRAPHY_FEATURE_ENABLED_KEY = 'coblocks_typography_controls_enabled';
+export const deprecatedBlocks = [ 'coblocks/click-to-tweet' ];
